requirements.txt
fastmcp
markdownify
readabilipy
httpx
pydantic
python-dotenv